split runtime into c modules
