+++ asn1crt/BitStream.c
@@ -1,15 +1,12 @@
+
+#include "BitStream.h"
+
 #include <string.h>
 #include <assert.h>
 #include <math.h>
 #include <float.h>
 
-#include "asn1crt.h"
-
-#ifndef INFINITY
-  #ifdef __GNUC__
-    #define INFINITY (__builtin_inf())
-  #endif
-#endif
+#include "asn1crt_util.h"
 
 static byte masks[] = { 0x80, 0x40, 0x20, 0x10, 0x08, 0x04, 0x02, 0x01 };
 static byte masksb[] = { 0x0, 0x1, 0x3, 0x7, 0xF, 0x1F, 0x3F, 0x7F, 0xFF };
@@ -24,13 +21,13 @@ static asn1SccUint32 masks2[] = { 0x0,
 /*   Bit Stream Functions                                                                      */
 /***********************************************************************************************/
 
-void BitStream_Init(BitStream* pBitStrm, byte* buf, long count) 
+void BitStream_Init(BitStream* pBitStrm, byte* buf, long count)
 {
     pBitStrm->count = count;
     pBitStrm->buf = buf;
     memset(pBitStrm->buf,0x0,(size_t)count);
     pBitStrm->currentByte = 0;
-    pBitStrm->currentBit=0; 
+    pBitStrm->currentBit=0;
 }
 
 void BitStream_AttachBuffer(BitStream* pBitStrm, unsigned char* buf, long count)
@@ -38,7 +35,7 @@ void BitStream_AttachBuffer(BitStream* pBitStrm, unsigned char* buf, long count)
     pBitStrm->count = count;
     pBitStrm->buf = buf;
     pBitStrm->currentByte = 0;
-    pBitStrm->currentBit=0; 
+    pBitStrm->currentBit=0;
 }
 
 asn1SccSint BitStream_GetLength(BitStream* pBitStrm)
@@ -50,11 +47,11 @@ asn1SccSint BitStream_GetLength(BitStream* pBitStrm)
 }
 
 
-void BitStream_AppendBitOne(BitStream* pBitStrm) 
+void BitStream_AppendBitOne(BitStream* pBitStrm)
 {
     pBitStrm->buf[pBitStrm->currentByte] |= masks[pBitStrm->currentBit];
 
-    if (pBitStrm->currentBit<7) 
+    if (pBitStrm->currentBit<7)
         pBitStrm->currentBit++;
     else {
         pBitStrm->currentBit=0;
@@ -63,9 +60,9 @@ void BitStream_AppendBitOne(BitStream* pBitStrm)
     assert(pBitStrm->currentByte*8+pBitStrm->currentBit<=pBitStrm->count*8);
 }
 
-void BitStream_AppendBitZero(BitStream* pBitStrm) 
+void BitStream_AppendBitZero(BitStream* pBitStrm)
 {
-    if (pBitStrm->currentBit<7) 
+    if (pBitStrm->currentBit<7)
         pBitStrm->currentBit++;
     else {
         pBitStrm->currentBit=0;
@@ -74,14 +71,14 @@ void BitStream_AppendBitZero(BitStream* pBitStrm)
     assert(pBitStrm->currentByte*8+pBitStrm->currentBit<=pBitStrm->count*8);
 }
 
-void BitStream_AppendNBitZero(BitStream* pBitStrm, int nbits) 
+void BitStream_AppendNBitZero(BitStream* pBitStrm, int nbits)
 {
     int totalBits = pBitStrm->currentBit + nbits;
     pBitStrm->currentBit = totalBits % 8;
     pBitStrm->currentByte += totalBits / 8;
 }
 
-void BitStream_AppendNBitOne(BitStream* pBitStrm, int nbits) 
+void BitStream_AppendNBitOne(BitStream* pBitStrm, int nbits)
 {
     int i;
 
@@ -110,12 +107,12 @@ void BitStream_AppendBits(BitStream* pBitStrm, const byte* srcBuffer, int nbits)
     }
 }
 
-void BitStream_AppendBit(BitStream* pBitStrm, flag v) 
+void BitStream_AppendBit(BitStream* pBitStrm, flag v)
 {
-    if (v) 
+    if (v)
         pBitStrm->buf[pBitStrm->currentByte] |= masks[pBitStrm->currentBit];
 
-    if (pBitStrm->currentBit<7) 
+    if (pBitStrm->currentBit<7)
         pBitStrm->currentBit++;
     else {
         pBitStrm->currentBit=0;
@@ -125,11 +122,11 @@ void BitStream_AppendBit(BitStream* pBitStrm, flag v)
 }
 
 
-flag BitStream_ReadBit(BitStream* pBitStrm, flag* v) 
+flag BitStream_ReadBit(BitStream* pBitStrm, flag* v)
 {
     *v = pBitStrm->buf[pBitStrm->currentByte] & masks[pBitStrm->currentBit];
-    
-    if (pBitStrm->currentBit<7) 
+
+    if (pBitStrm->currentBit<7)
         pBitStrm->currentBit++;
     else {
         pBitStrm->currentBit=0;
@@ -138,7 +135,7 @@ flag BitStream_ReadBit(BitStream* pBitStrm, flag* v)
     return pBitStrm->currentByte*8+pBitStrm->currentBit<=pBitStrm->count*8;
 }
 
-void BitStream_AppendByte(BitStream* pBitStrm, byte v, flag negate) 
+void BitStream_AppendByte(BitStream* pBitStrm, byte v, flag negate)
 {
     int cb = pBitStrm->currentBit;
     int ncb = 8 - cb;
@@ -153,7 +150,7 @@ void BitStream_AppendByte(BitStream* pBitStrm, byte v, flag negate)
 
 }
 
-void BitStream_AppendByte0(BitStream* pBitStrm, byte v) 
+void BitStream_AppendByte0(BitStream* pBitStrm, byte v)
 {
     int cb = pBitStrm->currentBit;
     int ncb = 8 - cb;
@@ -168,7 +165,7 @@ void BitStream_AppendByte0(BitStream* pBitStrm, byte v)
 }
 
 
-flag BitStream_ReadByte(BitStream* pBitStrm, byte* v) 
+flag BitStream_ReadByte(BitStream* pBitStrm, byte* v)
 {
     int cb = pBitStrm->currentBit;
     int ncb = 8 - pBitStrm->currentBit;
@@ -202,7 +199,7 @@ flag BitStream_ReadBits(BitStream* pBitStrm, byte* BuffToWrite, int nbits)
 }
 
 /* nbits 1..7*/
-void BitStream_AppendPartialByte(BitStream* pBitStrm, byte v, byte nbits, flag negate) 
+void BitStream_AppendPartialByte(BitStream* pBitStrm, byte v, byte nbits, flag negate)
 {
     int cb = pBitStrm->currentBit;
     int totalBits = cb+nbits;
@@ -217,7 +214,7 @@ void BitStream_AppendPartialByte(BitStream* pBitStrm, byte v, byte nbits, flag n
             pBitStrm->currentBit=0;
             pBitStrm->currentByte++;
         }
-    } 
+    }
     else {
         totalBitsForNextByte = totalBits - 8;
         pBitStrm->buf[pBitStrm->currentByte++] |= (byte)(v >> totalBitsForNextByte);
@@ -229,7 +226,7 @@ void BitStream_AppendPartialByte(BitStream* pBitStrm, byte v, byte nbits, flag n
 }
 
 /* nbits 1..7*/
-flag BitStream_ReadPartialByte(BitStream* pBitStrm, byte *v, byte nbits) 
+flag BitStream_ReadPartialByte(BitStream* pBitStrm, byte *v, byte nbits)
 {
     int cb = pBitStrm->currentBit;
     int totalBits = cb+nbits;
@@ -242,7 +239,7 @@ flag BitStream_ReadPartialByte(BitStream* pBitStrm, byte *v, byte nbits)
             pBitStrm->currentBit=0;
             pBitStrm->currentByte++;
         }
-    } 
+    }
     else {
         totalBitsForNextByte = totalBits - 8;
         *v = (byte)(pBitStrm->buf[pBitStrm->currentByte++] << totalBitsForNextByte);
@@ -313,7 +310,7 @@ static void BitStream_EncodeNonNegativeInteger32Neg(BitStream* pBitStrm,
 
 static flag BitStream_DecodeNonNegativeInteger32Neg(BitStream* pBitStrm,
                                                     asn1SccUint32* v,
-                                                    int nBits) 
+                                                    int nBits)
 {
     byte b;
     *v = 0;
@@ -324,7 +321,7 @@ static flag BitStream_DecodeNonNegativeInteger32Neg(BitStream* pBitStrm,
         *v|= b;
         nBits -=8;
     }
-    if (nBits) 
+    if (nBits)
     {
         *v<<=nBits;
         if (!BitStream_ReadPartialByte(pBitStrm, &b, (byte)nBits))
@@ -337,7 +334,7 @@ static flag BitStream_DecodeNonNegativeInteger32Neg(BitStream* pBitStrm,
 
 
 
-void BitStream_EncodeNonNegativeInteger(BitStream* pBitStrm, asn1SccUint v) 
+void BitStream_EncodeNonNegativeInteger(BitStream* pBitStrm, asn1SccUint v)
 {
 
 #if WORD_SIZE==8
@@ -347,7 +344,7 @@ void BitStream_EncodeNonNegativeInteger(BitStream* pBitStrm, asn1SccUint v)
         asn1SccUint32 hi = (asn1SccUint32)(v>>32);
         asn1SccUint32 lo = (asn1SccUint32)v;
         int nBits;
-        BitStream_EncodeNonNegativeInteger32Neg(pBitStrm, hi, 0); 
+        BitStream_EncodeNonNegativeInteger32Neg(pBitStrm, hi, 0);
 
         nBits = GetNumberOfBitsForNonNegativeInteger(lo);
         BitStream_AppendNBitZero(pBitStrm, 32-nBits);
@@ -385,7 +382,7 @@ flag BitStream_DecodeNonNegativeInteger(BitStream* pBitStrm, asn1SccUint* v, int
 }
 
 
-void BitStream_EncodeNonNegativeIntegerNeg(BitStream* pBitStrm, asn1SccUint v, flag negate) 
+void BitStream_EncodeNonNegativeIntegerNeg(BitStream* pBitStrm, asn1SccUint v, flag negate)
 {
 #if WORD_SIZE==8
     if (v<0x100000000LL)
@@ -408,94 +405,7 @@ void BitStream_EncodeNonNegativeIntegerNeg(BitStream* pBitStrm, asn1SccUint v, f
 #endif
 }
 
-static int GetNumberOfBitsForNonNegativeInteger32(asn1SccUint32 v) 
-{
-    int ret=0;
-
-    if (v<0x100) {
-        ret = 0;
-    } else if (v<0x10000) {
-        ret = 8;
-        v >>= 8;
-    } else if (v<0x1000000) {
-        ret = 16;
-        v >>= 16;
-    } else {
-        ret = 24;
-        v >>= 24;
-    }
-    while( v>0 ) {
-        v >>= 1;
-        ret++;
-    }
-    return ret;
-}
-
-int GetNumberOfBitsForNonNegativeInteger(asn1SccUint v) 
-{
-#if WORD_SIZE==8
-    if (v<0x100000000LL)
-        return GetNumberOfBitsForNonNegativeInteger32((asn1SccUint32)v);
-    else {
-        asn1SccUint32 hi = (asn1SccUint32)(v>>32);
-        return 32 + GetNumberOfBitsForNonNegativeInteger32(hi);
-    }
-#else
-    return GetNumberOfBitsForNonNegativeInteger32(v);
-#endif
-}
-
-int GetLengthInBytesOfUInt(asn1SccUint64 v);
-int GetLengthInBytesOfUInt(asn1SccUint64 v)
-{
-    int ret=0;
-    asn1SccUint32 v32 = (asn1SccUint32)v;
-#if WORD_SIZE==8
-    if (v>0xFFFFFFFF) {
-        ret = 4;
-        v32 = (asn1SccUint32)(v>>32);
-    } 
-#endif
-    
-    if (v32<0x100)
-        return ret+1;
-    if (v32<0x10000)
-        return ret+2;
-    if (v32<0x1000000)
-        return ret+3;
-    return ret+4;
-}
-
-static int GetLengthSIntHelper(asn1SccUint v)
-{
-    int ret=0;
-    asn1SccUint32 v32 = (asn1SccUint32)v;
-#if WORD_SIZE==8
-    if (v>0x7FFFFFFF) {
-        ret = 4;
-        v32 = (asn1SccUint32)(v>>32);
-    } 
-#endif
-
-    if (v32<=0x7F)
-        return ret+1;
-    if (v32<=0x7FFF)
-        return ret+2;
-    if (v32<=0x7FFFFF)
-        return ret+3;
-    return ret+4;
-}
-
-int GetLengthInBytesOfSInt(asn1SccSint v); // prototype to disable warnings
-int GetLengthInBytesOfSInt(asn1SccSint v)
-{
-    if (v >= 0)
-        return GetLengthSIntHelper((asn1SccUint)v);
-
-    return GetLengthSIntHelper((asn1SccUint)(-v - 1));
-}
-
-
+/***************/
 
 void BitStream_EncodeConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint v, asn1SccSint min, asn1SccSint max)
 {
@@ -534,9 +444,9 @@ flag BitStream_DecodeConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint* v, 
     asn1SccUint uv;
     int nRangeBits;
     asn1SccUint range = (asn1SccUint)(max-min);
-    
+
     ASSERT_OR_RETURN_FALSE(min<=max);
-    
+
 
     *v=0;
     if(!range) {
@@ -580,13 +490,13 @@ flag BitStream_DecodeConstraintPosWholeNumber(BitStream* pBitStrm, asn1SccUint* 
     return FALSE;
 }
 
-     
+
 void BitStream_EncodeSemiConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint v, asn1SccSint min)
 {
     int nBytes;
     assert(v>=min);
     nBytes = GetLengthInBytesOfUInt((asn1SccUint)(v - min));
-    
+
     /* encode length */
     BitStream_EncodeConstraintWholeNumber(pBitStrm, nBytes, 0, 255); /*8 bits, first bit is always 0*/
     /* put required zeros*/
@@ -648,7 +558,7 @@ flag BitStream_DecodeSemiConstraintPosWholeNumber(BitStream* pBitStrm, asn1SccUi
 void BitStream_EncodeUnConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint v)
 {
     int nBytes = GetLengthInBytesOfSInt(v);
-    
+
     /* encode length */
     BitStream_EncodeConstraintWholeNumber(pBitStrm, nBytes, 0, 255); /*8 bits, first bit is always 0*/
 
@@ -668,7 +578,7 @@ flag BitStream_DecodeUnConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint* v
     int i;
     flag valIsNegative=FALSE;
     *v=0;
-    
+
 
     if (!BitStream_DecodeConstraintWholeNumber(pBitStrm, &nBytes, 0, 255))
         return FALSE;
@@ -701,16 +611,16 @@ flag BitStream_DecodeUnConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint* v
             part 2 is 1 or more byte for exponent
             part 3 is 3 or more byte for mantissa (N)
 
-            First byte        
+            First byte
             S :0-->+, S:1-->-1
             Base will be always be 2 (implied by 6th and 5th bit which are zero)
             ab: F  (0..3)
-            cd:00 --> 1 byte for exponent as 2's complement 
-            cd:01 --> 2 byte for exponent as 2's complement 
-            cd:10 --> 3 byte for exponent as 2's complement 
-            cd:11 --> 1 byte for encoding the length of the exponent, then the expoent 
+            cd:00 --> 1 byte for exponent as 2's complement
+            cd:01 --> 2 byte for exponent as 2's complement
+            cd:10 --> 3 byte for exponent as 2's complement
+            cd:11 --> 1 byte for encoding the length of the exponent, then the expoent
 
-             8 7 6 5 4 3 2 1 
+             8 7 6 5 4 3 2 1
             +-+-+-+-+-+-+-+-+
             |1|S|0|0|a|b|c|d|
             +-+-+-+-+-+-+-+-+
@@ -726,13 +636,13 @@ void BitStream_EncodeReal(BitStream* pBitStrm, double v)
     asn1SccUint64 mantissa;
 
 
-    if (v==0.0) 
+    if (v==0.0)
     {
         BitStream_EncodeConstraintWholeNumber(pBitStrm, 0, 0, 0xFF);
         return;
     }
 
-    if (v == INFINITY ) 
+    if (v == INFINITY )
     {
         BitStream_EncodeConstraintWholeNumber(pBitStrm, 1, 0, 0xFF);
         BitStream_EncodeConstraintWholeNumber(pBitStrm, 0x40, 0, 0xFF);
@@ -783,7 +693,7 @@ void BitStream_EncodeReal(BitStream* pBitStrm, double v)
 
 }
 
-flag DecodeRealAsBinaryEncoding(BitStream* pBitStrm, int length, byte header, double* v);
+static flag DecodeRealAsBinaryEncoding(BitStream* pBitStrm, int length, byte header, double* v);
 
 flag BitStream_DecodeReal(BitStream* pBitStrm, double* v)
 {
@@ -817,7 +727,7 @@ flag BitStream_DecodeReal(BitStream* pBitStrm, double* v)
 }
 
 
-flag DecodeRealAsBinaryEncoding(BitStream* pBitStrm, int length, byte header, double* v)
+static flag DecodeRealAsBinaryEncoding(BitStream* pBitStrm, int length, byte header, double* v)
 {
     int sign=1;
     //int base=2;
@@ -880,77 +790,3 @@ flag DecodeRealAsBinaryEncoding(BitStream* pBitStrm, int length, byte header, do
 
 
 
-int GetCharIndex(char ch, byte Set[], int setLen)
-{
-    int i=0;
-    for(i=0; i<setLen; i++)
-        if (ch == Set[i])
-            return i;
-    return 0;
-}
-
-
-
-void ByteStream_Init(ByteStream* pStrm, byte* buf, long count) 
-{
-    pStrm->count = count;
-    pStrm->buf = buf;
-    memset(pStrm->buf,0x0,(size_t)count);
-    pStrm->currentByte = 0;
-    pStrm->EncodeWhiteSpace = FALSE;
-}
-
-void ByteStream_AttachBuffer(ByteStream* pStrm, unsigned char* buf, long count)
-{
-    pStrm->count = count;
-    pStrm->buf = buf;
-    pStrm->currentByte = 0;
-}
-
-asn1SccSint ByteStream_GetLength(ByteStream* pStrm)
-{
-    return pStrm->currentByte;
-}
-
-#if WORD_SIZE==8
-const asn1SccUint64 ber_aux[] = { 
-    0xFF,
-    0xFF00,
-    0xFF0000,
-    0xFF000000,
-    0xFF00000000ULL,
-    0xFF0000000000ULL,
-    0xFF000000000000ULL,
-    0xFF00000000000000ULL };
-#else
-const asn1SccUint32 ber_aux[] = {
-    0xFF,
-    0xFF00,
-    0xFF0000,
-    0xFF000000 };
-#endif
-
-
-
-asn1SccUint int2uint(asn1SccSint v) {
-    asn1SccUint ret = 0;
-    if (v < 0) {
-        ret = (asn1SccUint)(-v - 1);
-        ret = ~ret;
-    }
-    else {
-        ret = (asn1SccUint)v;
-    };
-    return ret;
-}
-
-asn1SccSint uint2int(asn1SccUint v, int uintSizeInBytes) {
-    int i;
-    asn1SccUint tmp = 0x80;
-    flag bIsNegative = (v & (tmp << ((uintSizeInBytes - 1) * 8)))>0;
-    if (!bIsNegative)
-        return (asn1SccSint)v;
-    for (i = WORD_SIZE - 1; i >= uintSizeInBytes; i--)
-        v |= ber_aux[i];
-    return -(asn1SccSint)(~v) - 1;
-}
