+++ asn1crt/BitStream.h
@@ -1,0 +1,66 @@
+
+#include "asn1crt_core.h"
+#include "asn1crt_real.h"
+
+typedef struct {
+	byte* buf;
+	long count;
+	long currentByte;
+	/* Next available bit for writting. Possible vallues 0..7, 0 is most significant bit of current byte*/
+	int currentBit;
+} BitStream;
+
+
+/* Bit strean functions */
+
+void BitStream_Init(BitStream* pBitStrm, unsigned char* buf, long count);
+void BitStream_AttachBuffer(BitStream* pBitStrm, unsigned char* buf, long count);
+void BitStream_AppendBit(BitStream* pBitStrm, flag v);
+void BitStream_AppendBits(BitStream* pBitStrm, const byte* srcBuffer, int nBitsToWrite);
+void BitStream_AppendByte(BitStream* pBitStrm, byte v, flag negate);
+void BitStream_AppendByte0(BitStream* pBitStrm, byte v);
+
+
+asn1SccSint BitStream_GetLength(BitStream* pBitStrm);
+void BitStream_AppendBitOne(BitStream* pBitStrm);
+void BitStream_AppendBitZero(BitStream* pBitStrm);
+flag BitStream_ReadBit(BitStream* pBitStrm, flag* v);
+flag BitStream_ReadBits(BitStream* pBitStrm, byte* BuffToWrite, int nBitsToRead);
+flag BitStream_ReadByte(BitStream* pBitStrm, byte* v);
+
+/* Integer functions */
+
+
+void BitStream_EncodeUnConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint v);
+void BitStream_EncodeSemiConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint v, asn1SccSint min);
+void BitStream_EncodeSemiConstraintPosWholeNumber(BitStream* pBitStrm, asn1SccUint v, asn1SccUint min);
+void BitStream_EncodeConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint v, asn1SccSint min, asn1SccSint max);
+void BitStream_EncodeConstraintPosWholeNumber(BitStream* pBitStrm, asn1SccUint v, asn1SccUint min, asn1SccUint max);
+
+flag BitStream_DecodeUnConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint* v);
+flag BitStream_DecodeSemiConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint* v, asn1SccSint min);
+flag BitStream_DecodeSemiConstraintPosWholeNumber(BitStream* pBitStrm, asn1SccUint* v, asn1SccUint min);
+flag BitStream_DecodeConstraintWholeNumber(BitStream* pBitStrm, asn1SccSint* v, asn1SccSint min, asn1SccSint max);
+flag BitStream_DecodeConstraintPosWholeNumber(BitStream* pBitStrm, asn1SccUint* v, asn1SccUint min, asn1SccUint max);
+
+
+void BitStream_EncodeReal(BitStream* pBitStrm, double v);
+flag BitStream_DecodeReal(BitStream* pBitStrm, double* v);
+
+
+/* Boolean Decode */
+
+flag BitStream_ReadBitPattern(BitStream* pBitStrm, const byte* patternToRead, int nBitsToRead, flag* pBoolValue);
+
+
+void BitStream_AppendNBitZero(BitStream* pBitStrm, int nbits);
+void BitStream_EncodeNonNegativeInteger(BitStream* pBitStrm, asn1SccUint v);
+void BitStream_AppendNBitOne(BitStream* pBitStrm, int nbits);
+void BitStream_EncodeNonNegativeIntegerNeg(BitStream* pBitStrm, asn1SccUint v, flag negate);
+flag BitStream_DecodeNonNegativeInteger(BitStream* pBitStrm, asn1SccUint* v, int nBits);
+flag BitStream_ReadPartialByte(BitStream* pBitStrm, byte *v, byte nbits);
+void BitStream_AppendPartialByte(BitStream* pBitStrm, byte v, byte nbits, flag negate);
+
+
+#define CHECK_BIT_STREAM(pBitStrm)	assert((pBitStrm)->currentByte*8+(pBitStrm)->currentBit<=(pBitStrm)->count*8)
+
