+++ asn1crt/real.c
@@ -1,7 +1,9 @@
 #include <assert.h>
 #include <float.h>
 #include <math.h>
-#include "asn1crt.h"
+
+#include "asn1crt_real.h"
+
 
 #ifndef FREXP_UNSUPPORTED
